Add BaseEditPage
- Move clickSaveButton, clickSaveDraftButton,
    clickSaveButtonErrorExpected, clickSaveDraftButtonErrorExpected
    to BaseEditPage
- Make FieldsEditPage inherited from BaseEditPage
